melhorias de qualidade de vida

+++ EstoquIN/View/Clientes.cs
@@ -127,11 +127,15 @@ namespace EstoquIN.View
 
         private void btnClienteExcluir_Click(object sender, EventArgs e)
         {
-            var t = context.DBclientes.Find((int)dataClientes.SelectedCells[0].Value);
-            context.DBclientes.Remove(t);
-            context.SaveChanges();
-            RefreshGrid();
-            ClearBoxes();
+            if (MessageBox.Show("TEM CERTEZA QUE DESEJA EXCLUIR?, ESTA AÇÃO É IRREVERSÍVEL", "EXCLUIR", MessageBoxButtons.OKCancel) == DialogResult.OK)
+
+            {
+                var t = context.DBclientes.Find((int)dataClientes.SelectedCells[0].Value);
+                context.DBclientes.Remove(t);
+                context.SaveChanges();
+                RefreshGrid();
+                ClearBoxes();
+            }
         }
     }
 }
