+++ EstoquIN/View/Fornecedores.cs
@@ -134,11 +134,15 @@ namespace EstoquIN.View
 
         private void btnFornecExcluir_Click(object sender, EventArgs e)
         {
-            var t = context.DBfornec.Find((int)dataFornec.SelectedCells[0].Value);
-            context.DBfornec.Remove(t);
-            context.SaveChanges();
-            RefreshGrid();
-            ClearBoxes();
+            if (MessageBox.Show("TEM CERTEZA QUE DESEJA EXCLUIR?, ESTA AÇÃO É IRREVERSÍVEL", "EXCLUIR", MessageBoxButtons.OKCancel) == DialogResult.OK)
+
+            {
+                var t = context.DBfornec.Find((int)dataFornec.SelectedCells[0].Value);
+                context.DBfornec.Remove(t);
+                context.SaveChanges();
+                RefreshGrid();
+                ClearBoxes();
+            }
         }
     }
 }
