+++ EstoquIN/View/Produtos.cs
@@ -113,11 +113,15 @@ namespace EstoquIN.View
 
         private void btnProdutoExcluir_Click(object sender, EventArgs e)
         {
-            var t = context.DBprodutos.Find((int)dataProduto.SelectedCells[0].Value);
-            context.DBprodutos.Remove(t);
-            context.SaveChanges();
-            RefreshGrid();
-            ClearBoxes();
+            if (MessageBox.Show("TEM CERTEZA QUE DESEJA EXCLUIR?, ESTA AÇÃO É IRREVERSÍVEL", "EXCLUIR", MessageBoxButtons.OKCancel) == DialogResult.OK)
+
+            {
+                var t = context.DBprodutos.Find((int)dataProduto.SelectedCells[0].Value);
+                context.DBprodutos.Remove(t);
+                context.SaveChanges();
+                RefreshGrid();
+                ClearBoxes();
+            }
         }
     }
 }
